refactor: create a utils and types and seperated it from component
