+++ frontend/src/container/GridGraphLayout/WidgetHeader/types.ts
@@ -1,0 +1,25 @@
+import { ReactNode } from 'react';
+
+import { MenuItemKeys } from './contants';
+
+export interface MenuItem {
+	key: string;
+	icon: ReactNode;
+	label: string;
+	isVisible: boolean;
+	disabled: boolean;
+	danger?: boolean;
+}
+
+export type TWidgetOptions =
+	| MenuItemKeys.View
+	| MenuItemKeys.Edit
+	| MenuItemKeys.Delete
+	| string;
+
+export type KeyMethodMappingProps<T extends TWidgetOptions> = {
+	[K in T]: {
+		key: TWidgetOptions;
+		method: VoidFunction | undefined;
+	};
+};
